requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Werkzeug==2.3.7
torch
transformers
datasets
anthropic
psycopg2-binary
python-dotenv
python-docx
openai-whisper
SpeechRecognition
numpy
sounddevice
scipy
openaipsycopg2-binary
